Fix interaction between ScopedkeyValueConfig and global field trials string
This patch fixes a problem for test programs that mix usage of
ScopedKeyValueConfig and the global field trial string.

In this case, tests that were using CallTest.

The solution is to check the global string when nothing has been explicitly added to a ScopedKeyValueConfig.

Bug: webrtc:13828
Change-Id: Ib89735670cfe93340ca0a8bac53f8a64a600ad66
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/256366
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#36294}