Remove crosscompile for  Perf Mac M1 Arm64 12
Since the Perf tester is actually compiling histogram code, it needs
to have it's own native OS gclient and GN config. The Builder has already
built and cached the MacOS arm64 binaries to be tested for M1.

I got confused and thought the tester was also compiling something for
the bot it shards to.

Bug: b/232347569
Change-Id: I0468945a04905d2d647626229af77c00d1f632fd
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/263200
Reviewed-by: Jeremy Leconte <jleconte@google.com>
Commit-Queue: Christoffer Jansson <jansson@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#36969}