Add missing dependency.
As mentioned in [1] declarations and definitions of the same symbol
should be part of the same library.

For some old code, this is not the case, and this can lead to hard to
debug linker errors like the ones from –warn-backrefs.

This CL adds the dependency to the defintion of call::Create() to
a target that uses it (and depends on the declaration from
call:call_interfaces).

In the future, call:call_interfaces should be removed entirely.

[1] - https://webrtc.googlesource.com/src/+/refs/heads/main/g3doc/style-guide/h-cc-pairs.md

Bug: None
Change-Id: I5f8fb6fa79815f1ff6b5199b9c682d7c9e73b616
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/276941
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Auto-Submit: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#38221}