De-flake NonSenderRttStats and make it faster to run on average.
It takes several seconds until we get an RTT measurement because that
requires RTCP packets to be received and those are not sent very often.

This CL makes the test faster on average by unblocking it as soon as
we see an RTT measurement (as opposed to always blocking for 10
seconds), this usually unblocks after around 5 seconds.

But to de-flake those rare instances where the test takes more than 10s
to run, the maximum timeout is extended to 20 seconds.

Patch Set 4: also fix use-of-uninitialized value.

Bug: webrtc:14981
Change-Id: Ieca94c90dfb52c3b17584a06660ff66c6462aa8b
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/296822
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Commit-Queue: Gustaf Ullberg <gustaf@webrtc.org>
Auto-Submit: Henrik Boström <hbos@webrtc.org>
Reviewed-by: Gustaf Ullberg <gustaf@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#39531}