Add legacy SVC test that all layers can be inactivated.
A larger version of this test was previously landed but got reverted
due to failures only happening on the importer bot (not on the CQ or
locally).

This is a smaller version of the test that does something we should
support: being able to inactive all encodings of a VP9 legacy SVC
stream.

Let's land and see if any issues are reproducible (expecting revert).

Bug: webrtc:15033
Change-Id: I88da1facf4ef05299f3392b86a0e3df029ebe264
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/299006
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Reviewed-by: Evan Shrubsole <eshr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#39684}