Change ChannelReceive::GetAudioFrameWithInfo to use new Converts method
Use the new Converts function added in webrtc-review.googlesource.com/c/src/+/320080. Later this will also be added to video.
This change is part of an effort to get Glass 2 Glass metrics. This particular change is not needed, but I intend to add this code to video, and thinks it's nice if the code for video and audio looks the same.

Bug: None
Change-Id: I04caff0dbef1cd4f391bbaa4f8bdee0e66043888
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/320281
Reviewed-by: Jakob Ivarsson‎ <jakobi@webrtc.org>
Commit-Queue: Olov Brändström <brandstrom@google.com>
Cr-Commit-Position: refs/heads/main@{#40753}